Compatible version for php5 and php7
This branch is based on version `0.1.*`,
it includes bin/checker. That was a requirement.
We used branch `0.2.*` for php7 apps.
The problem was dependencies in composer.
I solved it by adding Symfony ruleset to the
repository and by excluding dependencies
we don't need now.

This version works good on coma and orocrm.

Related [ch645]
Closes [ch640]